simplify(agent/runtime): tighten AgentRunLifecycle after extraction
Four cleanups surfaced by /simplify review:

- buildTerminalFlowResult moves to AgentFlowResult.ts where it belongs:
  it knows both workflow/toolUse result shapes, which are defined by the
  Zod schemas in that file.

- runFlowWithLifecycle sheds the redundant streamId, agentName, and
  category parameters; all three are already on ctx (ctx.streamId,
  ctx.config.agent, ctx.setting.agentCategory).  Callers pass ctx once
  and the function derives what it needs.

- category is removed from RunFlowLifecycleOptions for the same reason.

- getSdkErrorMessage(err) is extracted to a local sdkMsg in the catch
  block to avoid calling it twice.

Also fixes a missing executionRegistry import in executeAgent.ts (the
initial extract accidentally dropped it while executionRegistry.updateProgress
is still called in createRoundProgressCallback), and removes the dead
isWorkflowSetting and buildTerminalFlowResult imports that were carried
but never used.
